SDM-2639: Make code more testable.
